Profiling values are stored in a file First version on the greedy scheduler that reads profiling values from tables

+++ elastic_kernel.h
@@ -282,7 +282,7 @@ typedef struct{
 
 typedef struct {
 	int pairs[2];
-	int speedup;
+	double speedup;
 }t_co_speedup;
